\name{grasp.ascii}
\alias{grasp.ascii}
\title{Export data as an ascii grid file}
\description{Export data as ascii grid file that can be directly imported as a grid in Arcview for instance}
\usage{grasp.ascii(gr.Yi, var, resolution)}
\arguments{
	\item{gr.Yi}{Selected responses to be used in the form of a vector with column numbers}
	\item{var}{Dataframe to export. As default gr.predmat generated by grasp.pred()}
	\item{resolution}{Output grid resolution, 1000 as default}
}
\details{
    This function exports the result of prediction as a ascii grid file. It can then be directly imported in Arcview for instance.
    The filename is generated automatically and with pred, an underscore and the name of the variable Y you export, with the asc file extension.
}
\Note{If you export predictions of the same variable many times, the file is always overwritten.}
\author{
    Fabien Fivaz \email{fabien.fivaz@bluewin.ch}. Ported to R from GRASP \url{http://www.cscf.ch/grasp/} for S-Plus written by A. Lehmann,
    J.R. Leathwich and J. McC Overton. Look at \url{http://www.fivaz.ch/grasper/index.html} for details and update.
}
\seealso{
    \code{\link{grasp.pred}} and \code{\link{grasp.export}}
}
\keyword{models} \keyword{smooth} \keyword{regression}
