\name{grasp.export}
\alias{grasp.export}
\title{Export data as a point file}
\description{Export data as a point file readable with for instance Arcview}
\usage{grasp.export(var = gr.predmat)}
\arguments{
    \item{var}{Dataframe to export. As default: gr.predmat generated by \code{\link{grasp.pred}}}
}
\details{
    This function exports the result of prediction as a text file with index, coordinates and predictions with the first line containing the headers. 
    It can then be openend in Arcview. The filename is generated automatically and with pred, an underscore and the name of the variable Y you export, with the .txt 
    file extension.
}
\Note{If you export predictions of the same variable many times, the file is always overwritten.}
\author{
    Fabien Fivaz \email{fabien.fivaz@bluewin.ch}. Ported to R from GRASP \url{http://www.cscf.ch/grasp/} for S-Plus written by A. Lehmann, 
    J.R. Leathwich and J. McC Overton. Look at \url{http://www.fivaz.ch/grasper/index.html} for details and update.
}
\seealso{
    \code{\link{grasp.pred}} and \code{\link{grasp.ascii}}
}
\keyword{models} \keyword{regression}
